feat: add an api for other mods
Adds an api similar to GMD-API's.
The purpose of this API is to allow other mods to set/activate/get NONG
songs from Jukebox.
For now only adding songs and setting them as active is available from
the api.
At the moment, the only functionality this API offers is adding NONG
songs and activating a song. More functions can be added later.